chore: add selected icon
