+++ src/templates/workspace/snippets/listItemIcon.ts
@@ -1,26 +1,14 @@
 import * as pathLib from 'path';
 import * as vscode from 'vscode';
-import { t } from 'vscode-ext-localisation';
 
-export const listItemIcon = (
-  file: string,
-  label: string,
-  imgDarkFolderUri: vscode.Uri,
-  imgLightFolderUri: vscode.Uri
-) => {
-  const darkImgSrc = `${imgDarkFolderUri}${pathLib.sep}new-window.svg`;
-  const lightImgSrc = `${imgLightFolderUri}${pathLib.sep}new-window.svg`;
-  const tooltip = t('webViews.workspace.listItem.openNewWin', { label });
+export const listItemIcon = (imgDarkFolderUri: vscode.Uri, imgLightFolderUri: vscode.Uri) => {
+  const darkImgSrc = `${imgDarkFolderUri}${pathLib.sep}success.svg`;
+  const lightImgSrc = `${imgLightFolderUri}${pathLib.sep}success.svg`;
 
   return `
-        <span class="list__icon" data-file="${file}" data-type="new-window" title="${tooltip}">
-          <img alt="${t(
-            'webViews.workspace.listItem.iconAlt'
-          )}" data-theme="dark" src="${darkImgSrc}" />
-          <img alt="${t(
-            'webViews.workspace.listItem.iconAlt'
-          )}" data-theme="light" src="${lightImgSrc}" />
-        </span>
-      </span>
+    <span class="list__icon">
+      <img alt="" data-theme="dark" src="${darkImgSrc}" />
+      <img alt="" data-theme="light" src="${lightImgSrc}" />
+    </span>
   `;
 };
